Consolidate form views, only one file per type.
